Good enough for now
This is good enough for an initial deployment. It's not great, but the chief problems at this juncture is adding content.

Adding content, of course, will be my focus, but it would also be wise to refactor that content into something more sane in the codebase down the line. That's down the line, though.